test(lib-versioning): split the generate-changelog spec into focused suites
