+++ libs/versioning/src/flow/steps/generate-changelog.spec.ts
@@ -1,167 +1,5 @@
-import type { Logger } from '@hyperfrontend/logging'
-import type { Tree } from '@hyperfrontend/project-scope/vfs'
-import type { ChangelogEntry, ChangelogItem } from '../../changelog/models/entry'
-import type { ConventionalCommit } from '../../commits/models/conventional'
-import type { GitClient } from '../../git/factory'
-import type { Registry } from '../../registry/models/registry'
-import type { RepositoryConfig } from '../../repository/models'
-import type { FlowConfig, FlowContext, FlowState } from '../models/types'
-
-import { createGenerateChangelogStep, createWriteChangelogStep, GENERATE_CHANGELOG_STEP_ID } from './generate-changelog'
-
-function createMockLogger(): Logger {
-  return {
-    info: jest.fn(),
-    warn: jest.fn(),
-    error: jest.fn(),
-    debug: jest.fn(),
-    setLogLevel: jest.fn(),
-  } as unknown as Logger
-}
-
-interface MockTreeOptions {
-  files?: Record<string, string>
-}
-
-function createMockTree(options: MockTreeOptions = {}): Tree {
-  const files = new Map(Object.entries(options.files ?? {}))
-
-  return {
-    root: '/workspace',
-    read(path: string, encoding?: string) {
-      const content = files.get(path)
-      if (content === undefined) {
-        return null
-      }
-      return encoding ? content : Buffer.from(content)
-    },
-    write: jest.fn((path: string, content: string) => {
-      files.set(path, content)
-    }),
-    exists: (path: string) => files.has(path),
-    delete: jest.fn(),
-    rename: jest.fn(),
-    isFile: (path: string) => files.has(path),
-    children: () => [],
-    listChanges: () => [],
-  } as unknown as Tree
-}
-
-function createMockRegistry(): Registry {
-  return {
-    name: 'mock',
-    url: 'https://mock.registry.com',
-    async getLatestVersion() {
-      return null
-    },
-    async isVersionPublished() {
-      return false
-    },
-    async getPackageInfo() {
-      return null
-    },
-    async getVersionInfo() {
-      return null
-    },
-    async listVersions() {
-      return []
-    },
-  }
-}
-
-function createMockGitClient(): GitClient {
-  return {
-    cwd: '/workspace',
-    timeout: 30000,
-    getCommitLog: () => [],
-    getCommitsBetween: () => [],
-    getCommitsSince: () => [],
-    getCommit: () => null,
-    commitExists: () => true,
-    getTags: () => [],
-    getTag: () => null,
-    createTag: jest.fn(),
-    deleteTag: () => true,
-    tagExists: () => false,
-    getLatestTag: () => null,
-    getTagsForPackage: () => [],
-    pushTag: () => true,
-    createCommit: jest.fn(),
-    stage: jest.fn(),
-    unstage: () => true,
-    stageAll: jest.fn(),
-    amendCommit: jest.fn(),
-    createEmptyCommit: jest.fn(),
-    getHead: () => 'abc123',
-    getCurrentBranch: () => 'main',
-    hasStagedChanges: () => false,
-    hasUnstagedChanges: () => false,
-    hasUntrackedFiles: () => false,
-    getStatus: () => ({
-      clean: true,
-      entries: [],
-      staged: [],
-      unstaged: [],
-      untracked: [],
-    }),
-    isClean: () => true,
-    getRepositoryRoot: () => '/workspace',
-    getHeadHash: () => 'abc123',
-    getHeadShortHash: () => 'abc123',
-    getModifiedFiles: () => [],
-    getUntrackedFiles: () => [],
-    getStagedFiles: () => [],
-  } as unknown as GitClient
-}
-
-function createMockCommit(overrides: Partial<ConventionalCommit> = {}): ConventionalCommit {
-  return {
-    type: 'feat',
-    scope: [],
-    subject: 'test commit',
-    breaking: false,
-    breakingDescription: undefined,
-    body: undefined,
-    footers: [],
-    raw: 'feat: test commit',
-    ...overrides,
-  }
-}
-
-function createMockChangelogItem(description: string): ChangelogItem {
-  return { description, commits: [], references: [], breaking: false }
-}
-
-function createMockChangelogEntry(overrides: Partial<ChangelogEntry> = {}): ChangelogEntry {
-  return {
-    version: '1.0.0',
-    date: '2024-01-15',
-    unreleased: false,
-    sections: [
-      {
-        type: 'features' as const,
-        heading: 'Features',
-        items: [createMockChangelogItem('Feature 1')],
-      },
-    ],
-    ...overrides,
-  }
-}
-
-function createMockContext(state: Partial<FlowState> = {}, config: Partial<FlowConfig> = {}): FlowContext {
-  return {
-    workspaceRoot: '/workspace',
-    projectName: 'lib-test',
-    projectRoot: '/workspace/libs/test',
-    packageName: '@test/pkg',
-    tree: createMockTree(),
-    registry: createMockRegistry(),
-    git: createMockGitClient(),
-    logger: createMockLogger(),
-    config: { preset: 'conventional', ...config },
-    state: { ...state },
-  }
-}
+import { createMockCommit, createMockContext } from './__test-utils__/generate-changelog-mocks'
+import { createGenerateChangelogStep, GENERATE_CHANGELOG_STEP_ID } from './generate-changelog'
 
 describe('Generate Changelog Step', () => {
   describe('createGenerateChangelogStep', () => {
@@ -677,1330 +515,6 @@ describe('Generate Changelog Step', () => {
 
       expect(result.message).toMatch(/\d+ section\(s\)/)
       expect(result.message).toMatch(/\d+ commit\(s\)/)
-    })
-  })
-
-  describe('execute - compare URL generation', () => {
-    // Note: Compare URLs now use commit hashes only (not tags)
-
-    const createGitHubConfig = (): RepositoryConfig => ({
-      platform: 'github',
-      baseUrl: 'https://github.com/owner/repo',
-    })
-
-    const createGitLabConfig = (): RepositoryConfig => ({
-      platform: 'gitlab',
-      baseUrl: 'https://gitlab.com/group/project',
-    })
-
-    const createBitbucketConfig = (): RepositoryConfig => ({
-      platform: 'bitbucket',
-      baseUrl: 'https://bitbucket.org/owner/repo',
-    })
-
-    it('does not include compareUrl when repositoryConfig is not present', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        effectiveBaseCommit: 'def456789',
-        commits: [createMockCommit({ type: 'feat', subject: 'new feature' })],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(result.stateUpdates?.changelogEntry.compareUrl).toBeUndefined()
-    })
-
-    it('does not include compareUrl when effectiveBaseCommit is not present', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        repositoryConfig: createGitHubConfig(),
-        commits: [createMockCommit({ type: 'feat', subject: 'new feature' })],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(result.stateUpdates?.changelogEntry.compareUrl).toBeUndefined()
-    })
-
-    it('does not include compareUrl when effectiveBaseCommit is null', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        repositoryConfig: createGitHubConfig(),
-        effectiveBaseCommit: null,
-        commits: [createMockCommit({ type: 'feat', subject: 'new feature' })],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(result.stateUpdates?.changelogEntry.compareUrl).toBeUndefined()
-    })
-
-    it('logs info when publishedCommit exists but effectiveBaseCommit is null (fallback mode)', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        repositoryConfig: createGitHubConfig(),
-        publishedCommit: 'orphaned123',
-        effectiveBaseCommit: null,
-        commits: [createMockCommit({ type: 'feat', subject: 'new feature' })],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(result.stateUpdates?.changelogEntry.compareUrl).toBeUndefined()
-      expect(ctx.logger.info).toHaveBeenCalledWith('Compare URL omitted: published commit not in current history')
-    })
-
-    it('generates GitHub compareUrl using commit hashes when effectiveBaseCommit exists', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '1.1.0',
-        bumpType: 'minor',
-        repositoryConfig: createGitHubConfig(),
-        effectiveBaseCommit: 'def456789',
-        commits: [createMockCommit({ type: 'feat', subject: 'new feature' })],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(result.stateUpdates?.changelogEntry.compareUrl).toBe('https://github.com/owner/repo/compare/def456789...abc123')
-    })
-
-    it('generates GitLab compareUrl with /-/ prefix using commit hashes', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '2.0.0',
-        bumpType: 'major',
-        repositoryConfig: createGitLabConfig(),
-        effectiveBaseCommit: 'def456789',
-        commits: [createMockCommit({ type: 'feat', subject: 'breaking change', breaking: true })],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(result.stateUpdates?.changelogEntry.compareUrl).toBe('https://gitlab.com/group/project/-/compare/def456789...abc123')
-    })
-
-    it('generates Bitbucket compareUrl with reversed order and two dots using commit hashes', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '1.0.1',
-        bumpType: 'patch',
-        repositoryConfig: createBitbucketConfig(),
-        effectiveBaseCommit: 'def456789',
-        commits: [createMockCommit({ type: 'fix', subject: 'bug fix' })],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(result.stateUpdates?.changelogEntry.compareUrl).toBe('https://bitbucket.org/owner/repo/compare/abc123..def456789')
-    })
-
-    it('handles initial release path with repositoryConfig but no commits', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '0.1.0',
-        bumpType: 'minor',
-        repositoryConfig: createGitHubConfig(),
-        commits: [],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(result.message).toContain('initial release')
-      expect(result.stateUpdates?.changelogEntry.compareUrl).toBeUndefined()
-    })
-
-    it('generates compareUrl for version reset scenario (empty commits but has effectiveBaseCommit)', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'major',
-        repositoryConfig: createGitHubConfig(),
-        effectiveBaseCommit: 'def456789',
-        commits: [],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(result.stateUpdates?.changelogEntry.compareUrl).toBe('https://github.com/owner/repo/compare/def456789...abc123')
-    })
-
-    it('uses custom formatter when platform is custom', async () => {
-      const step = createGenerateChangelogStep()
-      const customConfig: RepositoryConfig = {
-        platform: 'custom',
-        baseUrl: 'https://my-git.internal/repo',
-        formatCompareUrl: (from, to) => `https://my-git.internal/diff/${from}/${to}`,
-      }
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        repositoryConfig: customConfig,
-        effectiveBaseCommit: 'def456789',
-        commits: [createMockCommit({ type: 'feat', subject: 'feature' })],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.stateUpdates?.changelogEntry.compareUrl).toBe('https://my-git.internal/diff/def456789/abc123')
-    })
-
-    it('logs debug message with abbreviated commit hashes when generating compareUrl', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        repositoryConfig: createGitHubConfig(),
-        effectiveBaseCommit: 'def456789abcdef',
-        commits: [createMockCommit({ type: 'feat', subject: 'feature' })],
-      })
-
-      await step.execute(ctx)
-
-      expect(ctx.logger.debug).toHaveBeenCalledWith('Compare URL: def4567...abc123')
-    })
-  })
-})
-
-describe('Write Changelog Step', () => {
-  describe('createWriteChangelogStep', () => {
-    it('creates a step with correct ID and name', () => {
-      const step = createWriteChangelogStep()
-
-      expect(step.id).toBe('write-changelog')
-      expect(step.name).toBe('Write Changelog')
-    })
-
-    it('depends on generate-changelog step', () => {
-      const step = createWriteChangelogStep()
-
-      expect(step.dependsOn).toContain('generate-changelog')
-    })
-  })
-
-  describe('execute - skip conditions', () => {
-    it('skips when no nextVersion', async () => {
-      const step = createWriteChangelogStep()
-      const ctx = createMockContext({ bumpType: 'minor', changelogEntry: createMockChangelogEntry() })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('skipped')
-    })
-
-    it('skips when bumpType is none', async () => {
-      const step = createWriteChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'none',
-        changelogEntry: createMockChangelogEntry({ version: '1.0.0', sections: [] }),
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('skipped')
-    })
-
-    it('skips when no changelog entry', async () => {
-      const step = createWriteChangelogStep()
-      const ctx = createMockContext({ nextVersion: '1.0.0', bumpType: 'minor' })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('skipped')
-    })
-
-    it('skips when changelog disabled', async () => {
-      const step = createWriteChangelogStep()
-      const ctx = createMockContext(
-        {
-          nextVersion: '1.0.0',
-          bumpType: 'minor',
-          changelogEntry: createMockChangelogEntry({ version: '1.0.0', sections: [] }),
-        },
-        { skipChangelog: true }
-      )
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('skipped')
-    })
-  })
-
-  describe('execute - create new changelog', () => {
-    it('creates new CHANGELOG.md when none exists', async () => {
-      const step = createWriteChangelogStep()
-      const tree = createMockTree()
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '1.0.0',
-          bumpType: 'minor',
-          changelogEntry: createMockChangelogEntry({
-            version: '1.0.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('Initial feature')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(tree.write).toHaveBeenCalled()
-      expect(result.message).toContain('Created CHANGELOG.md')
-    })
-
-    it('adds changelog path to modified files', async () => {
-      const step = createWriteChangelogStep()
-      const tree = createMockTree()
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '1.0.0',
-          bumpType: 'minor',
-          changelogEntry: createMockChangelogEntry({ version: '1.0.0', sections: [] }),
-          modifiedFiles: [],
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.stateUpdates?.modifiedFiles).toContain('/workspace/libs/test/CHANGELOG.md')
-    })
-  })
-
-  describe('execute - update existing changelog', () => {
-    it('adds entry to existing CHANGELOG.md', async () => {
-      const existingChangelog = `# Changelog
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '0.2.0',
-          bumpType: 'minor',
-          changelogEntry: createMockChangelogEntry({
-            version: '0.2.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('New feature')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(tree.write).toHaveBeenCalled()
-      expect(result.message).toContain('Updated CHANGELOG.md')
-    })
-
-    it('preserves existing modified files when updating', async () => {
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': '# Changelog\n' },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '1.0.0',
-          bumpType: 'minor',
-          changelogEntry: createMockChangelogEntry({ version: '1.0.0', sections: [] }),
-          modifiedFiles: ['/workspace/libs/test/package.json'],
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.stateUpdates?.modifiedFiles).toContain('/workspace/libs/test/package.json')
-      expect(result.stateUpdates?.modifiedFiles).toContain('/workspace/libs/test/CHANGELOG.md')
-    })
-  })
-
-  describe('execute - error handling', () => {
-    it('handles read error gracefully and creates new changelog', async () => {
-      const step = createWriteChangelogStep()
-      const logger = createMockLogger()
-
-      const throwingTree = {
-        root: '/workspace',
-        read: jest.fn((path: string) => {
-          if (path.includes('CHANGELOG.md')) {
-            throw new Error('Read error')
-          }
-          return null
-        }),
-        write: jest.fn(),
-        exists: () => false,
-        delete: jest.fn(),
-        rename: jest.fn(),
-        isFile: () => false,
-        children: () => [],
-        listChanges: () => [],
-      } as unknown as Tree
-
-      const ctx: FlowContext = {
-        workspaceRoot: '/workspace',
-        projectName: 'lib-test',
-        projectRoot: '/workspace/libs/test',
-        packageName: '@test/pkg',
-        tree: throwingTree,
-        registry: createMockRegistry(),
-        git: createMockGitClient(),
-        logger,
-        config: { preset: 'conventional' },
-        state: {
-          nextVersion: '1.0.0',
-          bumpType: 'minor',
-          changelogEntry: createMockChangelogEntry({ version: '1.0.0', sections: [] }),
-        },
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(logger.debug).toHaveBeenCalledWith('No existing CHANGELOG.md found')
-      expect(result.message).toContain('Created CHANGELOG.md')
-    })
-  })
-
-  describe('execute - classification result', () => {
-    function createMockRawCommit(hash = 'abc123') {
-      return {
-        hash,
-        shortHash: hash.slice(0, 7),
-        authorName: 'Test',
-        authorEmail: 'test@test.com',
-        authorDate: '2024-01-15T10:00:00Z',
-        committerName: 'Test',
-        committerEmail: 'test@test.com',
-        commitDate: '2024-01-15T10:00:00Z',
-        subject: 'test commit',
-        body: '',
-        message: 'test commit',
-        parents: [],
-        refs: [],
-      }
-    }
-
-    function createMockSummary(overrides: Partial<{ total: number; included: number; excluded: number }> = {}) {
-      return {
-        total: overrides.total ?? 1,
-        included: overrides.included ?? 1,
-        excluded: overrides.excluded ?? 0,
-        bySource: {
-          'direct-scope': 0,
-          'direct-file': 0,
-          'unscoped-file': 0,
-          'indirect-dependency': 0,
-          'indirect-infra': 0,
-          'unscoped-global': 0,
-          excluded: 0,
-        },
-      }
-    }
-
-    function createClassifiedCommit(
-      overrides: Partial<{
-        type: string
-        scope: readonly string[]
-        subject: string
-        breaking: boolean
-        breakingDescription: string
-        source: 'direct-scope' | 'direct-file' | 'unscoped-file' | 'indirect-dependency' | 'indirect-infra'
-        include: boolean
-        preserveScope: boolean
-      }> = {}
-    ) {
-      const {
-        type = 'feat',
-        scope = [],
-        subject = 'test feature',
-        breaking = false,
-        breakingDescription,
-        source = 'direct-scope',
-        include = true,
-        preserveScope = source !== 'direct-scope',
-      } = overrides
-
-      return {
-        commit: {
-          type,
-          scope,
-          subject,
-          breaking,
-          breakingDescription,
-          body: undefined,
-          footers: [],
-          raw: `${type}${scope.length > 0 ? `(${scope.join(',')})` : ''}: ${subject}`,
-        },
-        raw: createMockRawCommit(),
-        source,
-        include,
-        preserveScope,
-      }
-    }
-
-    it('uses classificationResult when available', async () => {
-      const step = createGenerateChangelogStep()
-      const directCommit = createClassifiedCommit({
-        type: 'feat',
-        scope: ['lib-test'],
-        subject: 'add new feature',
-        source: 'direct-scope',
-      })
-
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        commits: [createMockCommit({ type: 'feat', scope: ['lib-test'], subject: 'add new feature' })],
-        classificationResult: {
-          commits: [directCommit],
-          included: [directCommit],
-          excluded: [],
-          summary: createMockSummary(),
-        },
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      const entry = result.stateUpdates?.changelogEntry
-      expect(entry.sections.length).toBeGreaterThan(0)
-    })
-
-    it('omits scope for direct-scope commits (redundant)', async () => {
-      const step = createGenerateChangelogStep()
-      const directCommit = createClassifiedCommit({
-        type: 'feat',
-        scope: ['lib-test'],
-        subject: 'add feature',
-        source: 'direct-scope',
-        preserveScope: false,
-      })
-
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        commits: [createMockCommit({ type: 'feat', scope: ['lib-test'], subject: 'add feature' })],
-        classificationResult: {
-          commits: [directCommit],
-          included: [directCommit],
-          excluded: [],
-          summary: createMockSummary(),
-        },
-      })
-
-      const result = await step.execute(ctx)
-
-      const entry = result.stateUpdates?.changelogEntry
-      const featuresSection = entry.sections.find((s: { type: string }) => s.type === 'features')
-      expect(featuresSection).toBeDefined()
-      expect(featuresSection.items[0].description).not.toContain('**lib-test:**')
-      expect(featuresSection.items[0].description).toBe('add feature')
-    })
-
-    it('preserves scope for direct-file commits (informative)', async () => {
-      const step = createGenerateChangelogStep()
-      const directFileCommit = createClassifiedCommit({
-        type: 'feat',
-        scope: ['lib-other'],
-        subject: 'cross-project feature',
-        source: 'direct-file',
-        preserveScope: true,
-      })
-
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        commits: [createMockCommit({ type: 'feat', scope: ['lib-other'], subject: 'cross-project feature' })],
-        classificationResult: {
-          commits: [directFileCommit],
-          included: [directFileCommit],
-          excluded: [],
-          summary: createMockSummary(),
-        },
-      })
-
-      const result = await step.execute(ctx)
-
-      const entry = result.stateUpdates?.changelogEntry
-      const featuresSection = entry.sections.find((s: { type: string }) => s.type === 'features')
-      expect(featuresSection).toBeDefined()
-      expect(featuresSection.items[0].description).toContain('**lib-other:**')
-    })
-
-    it('preserves scope for indirect-dependency commits', async () => {
-      const step = createGenerateChangelogStep()
-      const indirectCommit = createClassifiedCommit({
-        type: 'feat',
-        scope: ['lib-utils'],
-        subject: 'utility improvement',
-        source: 'indirect-dependency',
-        preserveScope: true,
-      })
-
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        commits: [createMockCommit({ type: 'feat', scope: ['lib-utils'], subject: 'utility improvement' })],
-        classificationResult: {
-          commits: [indirectCommit],
-          included: [indirectCommit],
-          excluded: [],
-          summary: createMockSummary(),
-        },
-      })
-
-      const result = await step.execute(ctx)
-
-      const entry = result.stateUpdates?.changelogEntry
-      const depSection = entry.sections.find((s: { heading: string }) => s.heading === 'Dependency Updates')
-      expect(depSection).toBeDefined()
-      expect(depSection.items[0].description).toContain('**lib-utils:**')
-    })
-
-    it('creates Dependency Updates section for indirect commits', async () => {
-      const step = createGenerateChangelogStep()
-      const directCommit = createClassifiedCommit({
-        type: 'feat',
-        subject: 'direct feature',
-        source: 'direct-scope',
-      })
-      const indirectCommit = createClassifiedCommit({
-        type: 'fix',
-        scope: ['lib-dep'],
-        subject: 'dependency fix',
-        source: 'indirect-dependency',
-        preserveScope: true,
-      })
-
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        commits: [
-          createMockCommit({ type: 'feat', subject: 'direct feature' }),
-          createMockCommit({ type: 'fix', scope: ['lib-dep'], subject: 'dependency fix' }),
-        ],
-        classificationResult: {
-          commits: [directCommit, indirectCommit],
-          included: [directCommit, indirectCommit],
-          excluded: [],
-          summary: createMockSummary({ total: 2, included: 2 }),
-        },
-      })
-
-      const result = await step.execute(ctx)
-
-      const entry = result.stateUpdates?.changelogEntry
-      const featuresSection = entry.sections.find((s: { type: string }) => s.type === 'features')
-      const depSection = entry.sections.find((s: { heading: string }) => s.heading === 'Dependency Updates')
-
-      expect(featuresSection).toBeDefined()
-      expect(depSection).toBeDefined()
-      expect(depSection.items[0].description).toContain('**lib-dep:**')
-    })
-
-    it('sets indirect flag on ChangelogItem for indirect commits', async () => {
-      const step = createGenerateChangelogStep()
-      const indirectCommit = createClassifiedCommit({
-        type: 'feat',
-        scope: ['lib-infra'],
-        subject: 'infra update',
-        source: 'indirect-infra',
-        preserveScope: true,
-      })
-
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        commits: [createMockCommit({ type: 'feat', scope: ['lib-infra'], subject: 'infra update' })],
-        classificationResult: {
-          commits: [indirectCommit],
-          included: [indirectCommit],
-          excluded: [],
-          summary: createMockSummary(),
-        },
-      })
-
-      const result = await step.execute(ctx)
-
-      const entry = result.stateUpdates?.changelogEntry
-      const depSection = entry.sections.find((s: { heading: string }) => s.heading === 'Dependency Updates')
-      expect(depSection).toBeDefined()
-      expect(depSection.items[0].indirect).toBe(true)
-      expect(depSection.items[0].source).toBe('indirect-infra')
-    })
-
-    it('handles breaking changes with classification', async () => {
-      const step = createGenerateChangelogStep()
-      const breakingCommit = createClassifiedCommit({
-        type: 'feat',
-        scope: ['lib-test'],
-        subject: 'major API change',
-        breaking: true,
-        breakingDescription: 'Removed deprecated method',
-        source: 'direct-scope',
-      })
-
-      const ctx = createMockContext({
-        nextVersion: '2.0.0',
-        bumpType: 'major',
-        commits: [
-          createMockCommit({
-            type: 'feat',
-            scope: ['lib-test'],
-            subject: 'major API change',
-            breaking: true,
-            breakingDescription: 'Removed deprecated method',
-          }),
-        ],
-        classificationResult: {
-          commits: [breakingCommit],
-          included: [breakingCommit],
-          excluded: [],
-          summary: createMockSummary(),
-        },
-      })
-
-      const result = await step.execute(ctx)
-
-      const entry = result.stateUpdates?.changelogEntry
-      const breakingSection = entry.sections.find((s: { type: string }) => s.type === 'breaking')
-      expect(breakingSection).toBeDefined()
-      expect(breakingSection.heading).toBe('Breaking Changes')
-      expect(breakingSection.items[0].breaking).toBe(true)
-    })
-
-    it('falls back to commits without classification for backward compatibility', async () => {
-      const step = createGenerateChangelogStep()
-      const ctx = createMockContext({
-        nextVersion: '1.0.0',
-        bumpType: 'minor',
-        commits: [createMockCommit({ type: 'feat', scope: ['other'], subject: 'fallback feature' })],
-      })
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      const entry = result.stateUpdates?.changelogEntry
-      const featuresSection = entry.sections.find((s: { type: string }) => s.type === 'features')
-      expect(featuresSection).toBeDefined()
-      expect(featuresSection.items[0].description).toContain('**other:**')
-    })
-  })
-
-  describe('execute - pending publication cleanup', () => {
-    it('removes stacked entries when isPendingPublication is true', async () => {
-      const existingChangelog = `# Changelog
-
-## [0.2.0] - 2024-01-10
-
-### Features
-
-- Unpublished feature
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '0.2.0',
-          bumpType: 'minor',
-          publishedVersion: '0.1.0',
-          isPendingPublication: true,
-          changelogEntry: createMockChangelogEntry({
-            version: '0.2.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('Updated feature')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      const writtenContent = (tree.write as jest.Mock).mock.calls[0][1] as string
-      expect(writtenContent).toContain('Updated feature')
-      expect(writtenContent).not.toContain('Unpublished feature')
-      expect(writtenContent).toContain('0.1.0')
-    })
-
-    it('removes multiple stacked entries when isPendingPublication is true', async () => {
-      const existingChangelog = `# Changelog
-
-## [0.3.0] - 2024-01-15
-
-### Features
-
-- Feature 0.3.0
-
-## [0.2.0] - 2024-01-10
-
-### Features
-
-- Feature 0.2.0
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const logger = createMockLogger()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '0.2.0',
-          bumpType: 'minor',
-          publishedVersion: '0.1.0',
-          isPendingPublication: true,
-          changelogEntry: createMockChangelogEntry({
-            version: '0.2.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('Correct feature')] }],
-          }),
-        }),
-        tree,
-        logger,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      const writtenContent = (tree.write as jest.Mock).mock.calls[0][1] as string
-      expect(writtenContent).toContain('Correct feature')
-      expect(writtenContent).not.toContain('Feature 0.3.0')
-      expect(writtenContent).not.toContain('Feature 0.2.0')
-      expect(writtenContent).toContain('0.1.0')
-      expect(logger.info).toHaveBeenCalledWith(expect.stringContaining('Removing stacked entries'))
-    })
-
-    it('preserves unreleased entry during cleanup', async () => {
-      const existingChangelog = `# Changelog
-
-## [Unreleased]
-
-### Features
-
-- Work in progress
-
-## [0.2.0] - 2024-01-10
-
-### Features
-
-- Unpublished feature
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '0.2.0',
-          bumpType: 'minor',
-          publishedVersion: '0.1.0',
-          isPendingPublication: true,
-          changelogEntry: createMockChangelogEntry({
-            version: '0.2.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('Updated feature')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      const writtenContent = (tree.write as jest.Mock).mock.calls[0][1] as string
-      expect(writtenContent).toContain('Unreleased')
-      expect(writtenContent).toContain('Work in progress')
-    })
-
-    it('does not remove entries when isPendingPublication is false', async () => {
-      const existingChangelog = `# Changelog
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '0.2.0',
-          bumpType: 'minor',
-          publishedVersion: '0.1.0',
-          isPendingPublication: false,
-          changelogEntry: createMockChangelogEntry({
-            version: '0.2.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('New feature')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      const writtenContent = (tree.write as jest.Mock).mock.calls[0][1] as string
-      expect(writtenContent).toContain('0.2.0')
-      expect(writtenContent).toContain('0.1.0')
-      expect(writtenContent).toContain('New feature')
-      expect(writtenContent).toContain('Initial release')
-    })
-
-    it('does not remove entries when isPendingPublication is undefined', async () => {
-      const existingChangelog = `# Changelog
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '0.2.0',
-          bumpType: 'minor',
-          changelogEntry: createMockChangelogEntry({
-            version: '0.2.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('New feature')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      const writtenContent = (tree.write as jest.Mock).mock.calls[0][1] as string
-      expect(writtenContent).toContain('0.2.0')
-      expect(writtenContent).toContain('0.1.0')
-    })
-
-    it('handles de-escalation scenario (version downgrade)', async () => {
-      const existingChangelog = `# Changelog
-
-## [0.2.0] - 2024-01-10
-
-### Features
-
-- Wrong version feature
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '0.1.1',
-          bumpType: 'patch',
-          publishedVersion: '0.1.0',
-          isPendingPublication: true,
-          changelogEntry: createMockChangelogEntry({
-            version: '0.1.1',
-            sections: [{ type: 'fixes' as const, heading: 'Bug Fixes', items: [createMockChangelogItem('Bug fix')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      const writtenContent = (tree.write as jest.Mock).mock.calls[0][1] as string
-      expect(writtenContent).toContain('0.1.1')
-      expect(writtenContent).not.toContain('0.2.0')
-      expect(writtenContent).toContain('Bug fix')
-      expect(writtenContent).not.toContain('Wrong version feature')
-    })
-
-    it('handles escalation scenario (version upgrade)', async () => {
-      const existingChangelog = `# Changelog
-
-## [0.1.1] - 2024-01-10
-
-### Bug Fixes
-
-- Wrong version fix
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '0.2.0',
-          bumpType: 'minor',
-          publishedVersion: '0.1.0',
-          isPendingPublication: true,
-          changelogEntry: createMockChangelogEntry({
-            version: '0.2.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('New feature')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      const writtenContent = (tree.write as jest.Mock).mock.calls[0][1] as string
-      expect(writtenContent).toContain('0.2.0')
-      expect(writtenContent).not.toContain('0.1.1')
-      expect(writtenContent).toContain('New feature')
-      expect(writtenContent).not.toContain('Wrong version fix')
-    })
-
-    it('does not fail when publishedVersion is missing during pending state', async () => {
-      const existingChangelog = `# Changelog
-
-## [0.2.0] - 2024-01-10
-
-### Features
-
-- Some feature
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '0.2.0',
-          bumpType: 'minor',
-          isPendingPublication: true,
-          changelogEntry: createMockChangelogEntry({
-            version: '0.2.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('Updated feature')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-    })
-
-    it('uses replaceExisting when adding entry in pending publication state', async () => {
-      const existingChangelog = `# Changelog
-
-## [0.2.0] - 2024-01-10
-
-### Features
-
-- Old content
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '0.2.0',
-          bumpType: 'minor',
-          publishedVersion: '0.1.0',
-          isPendingPublication: true,
-          changelogEntry: createMockChangelogEntry({
-            version: '0.2.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('New content')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      const writtenContent = (tree.write as jest.Mock).mock.calls[0][1] as string
-      expect(writtenContent).toContain('New content')
-      expect(writtenContent).not.toContain('Old content')
-      const matches = writtenContent.match(/##\s+\[?0\.2\.0\]?/g)
-      expect(matches).toHaveLength(1)
-    })
-  })
-
-  describe('execute - custom changelog filename', () => {
-    it('uses changelogFileName from config when creating new changelog', async () => {
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({ files: {} })
-      const ctx: FlowContext = {
-        ...createMockContext(
-          {
-            nextVersion: '1.0.0',
-            bumpType: 'minor',
-            changelogEntry: createMockChangelogEntry({ version: '1.0.0' }),
-          },
-          { changelogFileName: 'HISTORY.md' }
-        ),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(tree.write).toHaveBeenCalledWith('/workspace/libs/test/HISTORY.md', expect.any(String))
-      expect(result.stateUpdates?.modifiedFiles).toContain('/workspace/libs/test/HISTORY.md')
-      expect(result.message).toContain('Created HISTORY.md')
-    })
-
-    it('uses changelogFileName from config when updating existing changelog', async () => {
-      const existingChangelog = `# Changelog
-
-## [0.1.0] - 2024-01-01
-
-### Features
-
-- Initial release
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({
-        files: { '/workspace/libs/test/RELEASES.md': existingChangelog },
-      })
-      const ctx: FlowContext = {
-        ...createMockContext(
-          {
-            nextVersion: '0.2.0',
-            bumpType: 'minor',
-            changelogEntry: createMockChangelogEntry({
-              version: '0.2.0',
-              sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('New feature')] }],
-            }),
-          },
-          { changelogFileName: 'RELEASES.md' }
-        ),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(tree.write).toHaveBeenCalledWith('/workspace/libs/test/RELEASES.md', expect.any(String))
-      expect(result.message).toContain('Updated RELEASES.md')
-    })
-
-    it('defaults to CHANGELOG.md when changelogFileName not specified', async () => {
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({ files: {} })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '1.0.0',
-          bumpType: 'minor',
-          changelogEntry: createMockChangelogEntry({ version: '1.0.0' }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(tree.write).toHaveBeenCalledWith('/workspace/libs/test/CHANGELOG.md', expect.any(String))
-    })
-  })
-
-  describe('execute - backupChangelog feature', () => {
-    it('creates backup when backupChangelog enabled and changelog exists', async () => {
-      const existingChangelog = `# Changelog
-
-## [1.0.0] - 2025-01-01
-
-### Features
-
-- Existing feature
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({ files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog } })
-      const ctx: FlowContext = {
-        ...createMockContext(
-          {
-            nextVersion: '1.1.0',
-            bumpType: 'minor',
-            changelogEntry: createMockChangelogEntry({
-              version: '1.1.0',
-              sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('New feature')] }],
-            }),
-          },
-          { backupChangelog: true }
-        ),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(tree.rename).toHaveBeenCalledWith('/workspace/libs/test/CHANGELOG.md', '/workspace/libs/test/CHANGELOG.backup.md')
-    })
-
-    it('deletes backup after successful write', async () => {
-      const existingChangelog = `# Changelog
-
-## [1.0.0] - 2025-01-01
-
-### Features
-
-- Existing feature
-`
-      const step = createWriteChangelogStep()
-      const files: Record<string, string> = { '/workspace/libs/test/CHANGELOG.md': existingChangelog }
-      const tree = createMockTree({ files })
-      const originalExists = tree.exists.bind(tree)
-      tree.exists = jest.fn((path: string) => {
-        if (path === '/workspace/libs/test/CHANGELOG.backup.md') return true
-        return originalExists(path)
-      })
-      const originalRead = tree.read.bind(tree)
-      tree.read = jest.fn((path: string, encoding?: string) => {
-        if (path === '/workspace/libs/test/CHANGELOG.backup.md') {
-          return encoding ? existingChangelog : Buffer.from(existingChangelog)
-        }
-        return originalRead(path, encoding as BufferEncoding)
-      }) as typeof tree.read
-      const ctx: FlowContext = {
-        ...createMockContext(
-          {
-            nextVersion: '1.1.0',
-            bumpType: 'minor',
-            changelogEntry: createMockChangelogEntry({
-              version: '1.1.0',
-              sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('New feature')] }],
-            }),
-          },
-          { backupChangelog: true }
-        ),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(tree.delete).toHaveBeenCalledWith('/workspace/libs/test/CHANGELOG.backup.md')
-    })
-
-    it('does not create backup when backupChangelog disabled', async () => {
-      const existingChangelog = `# Changelog
-
-## [1.0.0] - 2025-01-01
-
-### Features
-
-- Existing feature
-`
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({ files: { '/workspace/libs/test/CHANGELOG.md': existingChangelog } })
-      const ctx: FlowContext = {
-        ...createMockContext({
-          nextVersion: '1.1.0',
-          bumpType: 'minor',
-          changelogEntry: createMockChangelogEntry({
-            version: '1.1.0',
-            sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('New feature')] }],
-          }),
-        }),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(tree.rename).not.toHaveBeenCalled()
-    })
-
-    it('does not create backup when no existing changelog', async () => {
-      const step = createWriteChangelogStep()
-      const tree = createMockTree({ files: {} })
-      const ctx: FlowContext = {
-        ...createMockContext(
-          {
-            nextVersion: '1.0.0',
-            bumpType: 'minor',
-            changelogEntry: createMockChangelogEntry({
-              version: '1.0.0',
-              sections: [{ type: 'features' as const, heading: 'Features', items: [createMockChangelogItem('Initial feature')] }],
-            }),
-          },
-          { backupChangelog: true }
-        ),
-        tree,
-      }
-
-      const result = await step.execute(ctx)
-
-      expect(result.status).toBe('success')
-      expect(tree.rename).not.toHaveBeenCalled()
-      expect(result.message).toContain('Created')
     })
   })
 })
